+++ v3/changelog.md
@@ -4,6 +4,9 @@
 
 ### Added
 
+- Add flag to (`--expose-internal-errors`) toggle whether to expose internal
+  errors. ([#759](https://github.com/hasura/v3-engine/pull/759))
+
 #### Aggregates of Array Relationships
 
 Aggregates of array relationships can now be defined by specifying an